http error handling responses

+++ app/Http/Controllers/API/ProductsController.php
@@ -13,7 +13,10 @@ class ProductsController extends Controller
     public function create(CreateProductRequest $request)
     {
         $productData = $request->validated();
-
+        if (Product::where('id', $productData["id"])->exists()) {
+            // User with the specified userId already exists
+            return response()->json(["result" => "Product already exists"], 409);
+        }
         $product = new Product;
         $product->fill($productData);
         $product->ip = request()->ip();
@@ -27,17 +30,26 @@ class ProductsController extends Controller
     {
         
         $product = Product::where('id', $id)->get();
+        if($product->isEmpty()){
+            return response()->json(["message" => "Product not found"], 404);
+        }
         return $this->collectionResponse($product, new ProductTransformer);
     }
     public function destroy($id){
-        $product = Product::where('id', $id);
+        $product = Product::where('id', $id)->get();
+        if($product->isEmpty()){
+            return response()->json(["message" => "Product not found"], 404);
+        }
         $product->delete();
         return response()->json(["result" => "ok"], 200);
     }
-    public function update(Request $request,$id){
-        $productData = $request->all();
+    public function update(CreateProductRequest $request,$id){
+        $productData = $request->validated();
 
         $product = Product::where('id', $id)->first();
+        if(!$product){
+            return response()->json(["message" => "Product not found"], 404);
+        }
         $product->fill($productData);
         $product->save();
         
@@ -54,7 +66,9 @@ class ProductsController extends Controller
             $userId = $request->query("content");
 
             $user = User_new::where('userId', $userId)->first();
-            
+            if(!$user){
+                return response()->json(["message" => "User not found"], 404);
+            }
             if($options == "userProducts"){
                 $productIds = $user->products;
                 $Userproducts = Product::whereIn('id', $productIds)->get();
@@ -66,6 +80,7 @@ class ProductsController extends Controller
                 $Userproducts = Product::whereIn('id', $productIds)->get();
                 return $this->collectionResponse($Userproducts, new ProductTransformer);
             }
+            return response()->json(["message" => "Invalid category"], 422);
         }
         
     }
